fix(macos): empty listings page — listPriceEur was Int, API sends Double
The API returns listPriceEur as a float (e.g., 488000.5 from cents/100).
Swift DTO declared Int?, causing JSONDecoder to fail on the entire
paginated response. Changed to Double? with Int() conversion in toDomain.

Also verified end-to-end alert flow: filter → crawl → 3 alerts created.
Checklist: 438/448 (97.8%).

+++ apps/macos/RealEstateIntel/Networking/APIListingResponse.swift
@@ -14,7 +14,7 @@ struct APIListingResponse: Codable {
     let postalCode: String?
     let districtNo: Int?
     let districtName: String?
-    let listPriceEur: Int?
+    let listPriceEur: Double?
     let livingAreaSqm: Double?
     let rooms: Int?
     let pricePerSqmEur: Double?
@@ -43,7 +43,7 @@ struct APIListingResponse: Codable {
             postalCode: postalCode,
             districtNo: districtNo,
             districtName: districtName,
-            listPriceEur: listPriceEur ?? 0,
+            listPriceEur: Int(listPriceEur ?? 0),
             livingAreaSqm: livingAreaSqm,
             rooms: rooms,
             pricePerSqmEur: pricePerSqmEur,
